backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and Celery
redis==5.0.1
celery==5.3.4

# HTML Parsing
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Report Export
weasyprint==60.1
openpyxl==3.1.2
jinja2==3.1.2
pillow==10.1.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Security (Phase 2)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utils
python-dateutil==2.8.2
httpx==0.25.2
